Change types used in macros

+++ template/template.cpp
@@ -19,11 +19,11 @@ struct pair_hash {
 };
 
 #define LOG(variable) cerr << #variable":\t" << (variable) << "\n"
-#define LOGCON(i, container) for(int (i) = 0; (i) < (container).size(); ++(i)) cerr << (i) << ":\t" << (container)[(i)] << "\n"
-#define REP(i, n) for (int i = 0; i < (n); ++i)
-#define REPS(i, r, n) for (int i = (r); i < (n); ++i)
-#define REPR(i, n) for(int i = (n); i >= 0; --i) // from n to 0
-#define REPRS(i, n, r) for(int i = (n); i >= (r); --i) // from n to r
+#define LOGCON(i, container) for(size_t (i) = 0; (i) < (container).size(); ++(i)) cerr << (i) << ":\t" << (container)[(i)] << "\n"
+#define REP(i, n) for (long long i = 0; i < (n); ++i)
+#define REPS(i, r, n) for (long long i = (r); i < (n); ++i)
+#define REPR(i, n) for(long long i = (n); i >= 0; --i) // from n to 0
+#define REPRS(i, n, r) for(long long i = (n); i >= (r); --i) // from n to r
 #define REPOBJ(itr, obj) for(auto itr = (obj).begin(); itr != (obj).end() ; ++itr)
 #define REPROBJ(itr, obj) for(auto itr = (obj).rbegin(), e = (obj).rend(); itr != e; ++itr)
 #define COUTB(x) cout << (x) << "\n"
